+++ NEWS.md
@@ -8,6 +8,7 @@ Adding the following new methods:
 - `clear_and_summarise_messages()`
 - `judge_and_choose_best_response()`
 - `set_budget()`
+- `export_messages_history` and `load_messages_history`
 
 
 # mini007 0.1.0